fix(executable) Use Integer.MAX_VALUE rather than Long.MAX_VALUE for external task lock duration as Long.MAX_VALUE causes exceptions when using Oracle (#41)

+++ core/src/main/java/org/camunda/bpm/scenario/impl/waitstate/ServiceTaskExecutable.java
@@ -32,7 +32,7 @@ public class ServiceTaskExecutable extends AbstractExternalTaskDelegate {
   }
 
   protected void fetchAndLock() {
-    getExternalTaskService().fetchAndLock(Integer.MAX_VALUE, WORKER_ID).topic(getDelegate().getTopicName(), Long.MAX_VALUE).execute();
+    getExternalTaskService().fetchAndLock(Integer.MAX_VALUE, WORKER_ID).topic(getDelegate().getTopicName(), Integer.MAX_VALUE).execute();
   }
 
   @Override
